Updated and added files
All files for prototype one added, removed duplicates and irrelevant files for this prototype version
