+++ gui/speech.py
@@ -1,0 +1,24 @@
+#-----------------------------------------------------------------------
+# File Title: Speech Class
+# File Description: Handles all speech recognition functionality.
+# Author: Ryan Partridge
+#-----------------------------------------------------------------------
+
+#-----------------------------------------------------------------------
+# Class Title: Speech()
+#-----------------------------------------------------------------------
+class Speech():
+  """
+  Used to handle spoken user queries and extract the right information
+  from the database, based on the users query.\n
+  Functions: ()
+  """
+  #-----------------------------------------------------------------------
+  # Function Title: __init__()
+  #-----------------------------------------------------------------------
+  def __init__(self):
+    pass
+
+  #-----------------------------------------------------------------------
+  # Function Title: ()
+  #-----------------------------------------------------------------------
