--- gui/templates/error-page.html
@@ -1,45 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="author" content="Ryan Partridge">
-  <link rel="stylesheet" id="main-stylesheet" href="{{ url_for('static',filename='css/styles.css') }}">
-  <link rel="stylesheet" id="adobe-fonts" href="https://use.typekit.net/lqz6wgg.css">
-  <link rel="stylesheet" id="bootstrap-css" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
-  <script src="https://kit.fontawesome.com/37559427be.js"></script>
-  <title>University of Lincoln: Open Day</title>
-</head>
-
-<body>
-  <div class="container-fluid">
-    <div class="row blue-bg spacing">
-      <div class="col-lg-4 logo align-self-center">
-         <img src="{{ url_for('static', filename='images/uol-logo-small.jpg') }}" alt="University of Lincoln Logo">
-      </div><!-- /.logo -->
-      <div class="col-lg-4"></div>
-      <div class="col-lg-4 title-text">
-        <h2>Open Day</h2>
-        <h2>11<span>th</span> February 2020</h2>
-      </div><!-- /.title-text -->
-    </div><!-- /.row -->
-
-    <div class="container-fluid white-bg margin-spacing vertical-spacing">
-      <div class="row">
-        <div class="col-lg-12 error-message">
-          <i class="far fa-dizzy"></i>
-          <p>Sorry, I've never been asked that before!</p>
-          <p>Please speak to <span>[academic_name]</span> at the <br><span>[stool_name]</span></p>
-          <p>Alternatively, please email us at <br><span>[email_address]</span></p>
-        </div><!-- /.question-title -->
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-8 return-btn">
-          <button id="back-btn" type="button" onclick="history.back();"><i class="far fa-arrow-alt-circle-left"></i> Go back and try a different question</button>
-        </div><!-- /.return-btn -->
-      </div><!-- /.row -->
-    </div> <!-- /.container-content -->
-  </div> <!-- /.container -->
-</body>
-</html>
