+++ gui/templates/error.html
@@ -1,0 +1,18 @@
+{% extends "layout.html" %}
+{% block content %}
+    <div class="container-fluid white-bg margin-spacing vertical-spacing">
+      <div class="row">
+        <div class="col-lg-12 error-message">
+          <i class="far fa-dizzy"></i>
+          <p>Sorry, I've never been asked that before!</p>
+          <p>Please speak to <span>[academic_name]</span> at the <br><span>[stool_name]</span></p>
+          <p>Alternatively, please email us at <br><span>[email_address]</span></p>
+        </div><!-- /.question-title -->
+      </div><!-- /.row -->
+      <div class="row justify-content-center">
+        <div class="col-lg-8 return-btn">
+          <a id="back-btn" href="{{ url_for('home') }}"><i class="far fa-arrow-alt-circle-left"></i> Go back and try a different question</a>
+        </div><!-- /.return-btn -->
+      </div><!-- /.row -->
+    </div> <!-- /.container-content -->
+{% endblock %}
