--- gui/templates/home-p1.html
@@ -1,49 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="author" content="Ryan Partridge">
-  <link rel="stylesheet" id="main-stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-  <link rel="stylesheet" id="adobe-fonts" href="https://use.typekit.net/lqz6wgg.css">
-  <link rel="stylesheet" id="bootstrap-css" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
-  <title>University of Lincoln: Open Day</title>
-</head>
-
-<body>
-  <div class="container-fluid">
-    <div class="row blue-bg spacing">
-      <div class="col-lg-4 logo align-self-center">
-        <img src="{{ url_for('static', filename='images/uol-logo-small.jpg') }}" alt="University of Lincoln Logo">
-      </div><!-- /.logo -->
-      <div class="col-lg-4"></div>
-      <div class="col-lg-4 title-text">
-        <h2>Open Day</h2>
-        <h2>11<span>th</span> February 2020</h2>
-      </div><!-- /.title-text -->
-    </div><!-- /.row -->
-
-    <div class="row vertical-spacing justify-content-center">
-      <div class="col-lg-5 mx-4 category first">
-        <a id="uni-btn" href="{{ url_for('questions') }}">University</a>
-      </div><!-- /.category -->
-      <div class="col-lg-5 mx-4 category">
-        <a id="det-btn" href="{{ url_for('questions') }}">Course Details</a>
-      </div><!-- /.category -->
-    </div><!-- /.row -->
-
-    <div class="row vertical-spacing justify-content-center">
-      <div class="col-lg-5 mx-4 category">
-        <a id="acc-btn" href="{{ url_for('questions') }}">Accommodation</a>
-      </div><!-- /.category -->
-      <div class="col-lg-5 mx-4 category">
-        <a id="oth-btn" href="{{ url_for('questions') }}">Other</a>
-      </div><!-- /.category -->
-    </div><!-- /.row -->
-  </div> <!-- /.container -->
-    <!-- Custom scripts -->
-  <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-  <script type="text/javascript" src="custom.js"></script>
-</body>
-</html>
