--- gui/templates/home-p2.html
@@ -1,124 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="author" content="Ryan Partridge">
-  <link rel="stylesheet" id="main-stylesheet" href="{{ url_for('static',     filename='css/styles.css') }}">
-  <link rel="stylesheet" id="adobe-fonts" href="https://use.typekit.net/lqz6wgg.css">
-  <link rel="stylesheet" id="bootstrap-css" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
-  <script src="https://kit.fontawesome.com/37559427be.js"></script>
-  <title>University of Lincoln: Open Day</title>
-</head>
-
-<body>
-  <div class="container-fluid">
-    <div class="row blue-bg spacing">
-      <div class="col-lg-4 logo align-self-center">
-         <img src="{{ url_for('static', filename='images/uol-logo-small.jpg') }}" alt="University of Lincoln Logo">
-      </div><!-- /.logo -->
-      <div class="col-lg-4"></div>
-      <div class="col-lg-4 title-text">
-        <h2>Open Day</h2>
-        <h2>11<span>th</span> February 2020</h2>
-      </div><!-- /.title-text -->
-    </div><!-- /.row -->
-
-    <div class="container-fluid margin-spacing">
-      <div class="row">
-        <div class="col-lg-11 search-title white-bg">
-          <input type="text" id="searchbar" name="searchbar" placeholder="What can I help you with today?"></input>
-          <hr>
-        </div><!-- /.search-title -->
-        <div class="col-lg-1 search-icon">
-          <button id="search-btn" type="button">
-            <i class="fas fa-search"></i>
-          </button>
-        </div><!-- /.search-icon -->
-      </div><!-- /.row -->
-    </div><!-- /.search-container -->
-
-    <div class="container-fluid white-bg margin-spacing vertical-spacing">
-      <div class="row">
-        <div class="col-lg-12 question-title">
-          <h3>Recently asked questions...</h3>
-        </div><!-- /.question-title -->
-      </div><!-- /.row -->
-  
-      <div class="row justify-content-center">
-        <div class="col-lg-10 question left">
-          <p>
-            <a href="##">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt, lorem ipsum sit.</a>
-          </p>
-        </div><!-- /.question -->
-        <div class="col-lg-1 question right">
-          <p>
-            <a href="##">></a>
-          </p>
-        </div>
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-10 question left">
-          <p>
-            <a href="##">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt, lorem ipsum sit.</a>
-          </p>
-        </div><!-- /.question -->
-        <div class="col-lg-1 question right">
-          <p>
-            <a href="##">></a>
-          </p>
-        </div>
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-10 question left">
-          <p>
-            <a href="##">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt, lorem ipsum sit.</a>
-          </p>
-        </div><!-- /.question -->
-        <div class="col-lg-1 question right">
-          <p>
-            <a href="##">></a>
-          </p>
-        </div>
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-10 question left">
-          <p>
-            <a href="##">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt, lorem ipsum sit.</a>
-          </p>
-        </div><!-- /.question -->
-        <div class="col-lg-1 question right">
-          <p>
-            <a href="##">></a>
-          </p>
-        </div>
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-10 question left">
-          <p>
-            <a href="##">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt, lorem ipsum sit.</a>
-          </p>
-        </div><!-- /.question -->
-        <div class="col-lg-1 question right">
-          <p>
-            <a href="##">></a>
-          </p>
-        </div>
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-10 question left">
-          <p>
-            <a href="##">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt, lorem ipsum sit.</a>
-          </p>
-        </div><!-- /.question -->
-        <div class="col-lg-1 question right">
-          <p>
-            <a href="##">></a>
-          </p>
-        </div>
-      </div><!-- /.row -->
-    </div><!-- /.question-container-->
-  </div> <!-- /.container -->
-</body>
-</html>
