+++ gui/templates/layout.html
@@ -1,22 +1,28 @@
 <!doctype html>
 <html lang="en">
 <head>
+  {% block head %}
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="author" content="Ryan Partridge">
-  <link rel="stylesheet" id="main-stylesheet" href="{{ url_for('static',     filename='css/styles.css') }}">
+  <link rel='icon' href="{{ url_for('static', filename='images/icon.jpg') }}" type='image/x-icon'>
+  <link rel="stylesheet" id="main-stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
   <link rel="stylesheet" id="adobe-fonts" href="https://use.typekit.net/lqz6wgg.css">
   <link rel="stylesheet" id="bootstrap-css" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
+  <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
   <script src="https://kit.fontawesome.com/37559427be.js"></script>
   <title>University of Lincoln: Open Day</title>
+  {% endblock %}
 </head>
 
 <body>
   <div class="container-fluid">
     <div class="row blue-bg spacing">
       <div class="col-lg-4 logo align-self-center">
-         <img src="{{ url_for('static', filename='images/uol-logo-small.jpg') }}" alt="University of Lincoln Logo">
+        <a href="{{ url_for('home') }}">
+          <img src="{{ url_for('static', filename='images/uol-logo-small.jpg') }}" alt="University of Lincoln Logo">
+        </a>
       </div><!-- /.logo -->
       <div class="col-lg-4"></div>
       <div class="col-lg-4 title-text">
@@ -24,19 +30,10 @@
         <h2>11<span>th</span> February 2020</h2>
       </div><!-- /.title-text -->
     </div><!-- /.row -->
-
-    <div class="container-fluid white-bg margin-spacing vertical-spacing">
-      <div class="row">
-        <div class="col-lg-12 proto-three question-title">
-          <h3>What can I help you with today?</h3>
-          <div class="speech-container">
-            <i class="fa fa-microphone" aria-hidden="true"></i>
-            <div class="pulse-ring"></div>
-          </div><!-- /.speech-container -->
-          <p>Listening...</p>
-        </div><!-- /.question-title -->
-      </div><!-- /.row -->
-    </div><!-- /.question-container -->
+    {% block content %}
+    {% endblock %}
   </div> <!-- /.container -->
-</body>
-</html>
+  <!-- Custom scripts -->
+  <script type="text/javascript" src="{{ url_for('static', filename='js/custom.js') }}"></script>
+  </body>
+  </html>
