+++ gui/templates/questions.html
@@ -1,0 +1,25 @@
+{% extends "layout.html" %}
+{% block content %}
+    <div class="container-fluid white-bg margin-spacing vertical-spacing">
+      <div class="row">
+        <div class="col-lg-12 question-title">
+          <h3>What would you like to know?</h3>
+        </div><!-- /.question-title -->
+      </div><!-- /.row -->
+  
+      {% for item in session.get('queries') %}
+      <div class="row justify-content-center">
+        <div class="col-lg-11 question">
+          <form action="#" method="POST">
+            <input id="question" name="query" type="submit" value="{{ item }}">
+          </form>
+        </div><!-- /.question -->
+      </div><!-- /.row -->
+      {% endfor %}
+      <div class="row justify-content-center">
+        <div class="col-lg-8 return-btn p-one">
+          <a id="back-btn" href="{{ url_for('home') }}"><i class="far fa-arrow-alt-circle-left"></i> Go back and try a different category</a>
+        </div><!-- /.return-btn -->
+      </div><!-- /.row -->
+    </div><!-- /.question-container-->
+{% endblock %}
