--- gui/templates/results-page.html
@@ -1,49 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="author" content="Ryan Partridge">
-  <link rel="stylesheet" id="main-stylesheet" href="{{ url_for('static',     filename='css/styles.css') }}">
-  <link rel="stylesheet" id="adobe-fonts" href="https://use.typekit.net/lqz6wgg.css">
-  <link rel="stylesheet" id="bootstrap-css" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
-  <script src="https://kit.fontawesome.com/37559427be.js"></script>
-  <title>University of Lincoln: Open Day</title>
-</head>
-
-<body>
-  <div class="container-fluid">
-    <div class="row blue-bg spacing">
-      <div class="col-lg-4 logo align-self-center">
-         <img src="{{ url_for('static', filename='images/uol-logo-small.jpg') }}" alt="University of Lincoln Logo">
-      </div><!-- /.logo -->
-      <div class="col-lg-4"></div>
-      <div class="col-lg-4 title-text">
-        <h2>Open Day</h2>
-        <h2>11<span>th</span> February 2020</h2>
-      </div><!-- /.title-text -->
-    </div><!-- /.row -->
-
-    <div class="container-fluid white-bg margin-spacing vertical-spacing">
-      <div class="row">
-        <div class="col-lg-12 results-title">
-          <h2>When can I apply for accommodation?</h2>
-        </div><!-- /.results-title -->
-      </div><!-- /.row -->
-      <div class="row">
-        <div class="col-lg-12 results-description">
-          <p>From 2nd March the application system is open to any applicant who has accepted their course offer via UCAS, as either their first choice or insurance choice or if you awaiting an offer due to portfolio review, an interview or audition.</p>
-          <p>To apply you need to have accepted your offer at least 24 hours or one working day before you wish to submit an application. You will need your University of Lincoln applicant ID. You will have received this in your academic offer and it will also be included in an email which will be sent to you when you are able to apply for your accommodation.</p>
-          <p>*If you have deferred your place then you will need to reapply for accommodation from the 2nd March, in your year of entry. We will not allocate accommodation earlier to unconditional or deferred status applicants.</p>
-        </div><!-- /.results-description -->
-      </div><!-- /.row -->
-      <div class="row justify-content-center">
-        <div class="col-lg-8 return-btn">
-           <button id="back-btn" type="button" onclick="history.back();"><i class="far fa-arrow-alt-circle-left"></i> Go back and try a different question</button>
-        </div><!-- /.return-btn -->
-      </div><!-- /.row -->
-    </div> <!-- /.container-content -->
-  </div> <!-- /.container -->
-</body>
-</html>
