--- questions.csv
@@ -1,26 +1,0 @@
-What are the entry requirements?,BBC or 112 UCAS points,Course Overview
-How is the course accredited?,The course is accredited by The British Computer Society,Course Overview
-Can I do a placement?,The course can be taken with an optional sandwich placement. If students choose the sandwich placement they take a year out between second and third years to complete the placement.,
-How much are the tuition fees?,"For full time UK students each level is £9,250 for international full time students £15,900 per level.",University
-Can I do a Masters?,Students can apply for MComp Computer Science with an integrated Masters course.,
-What modules will I study?,"First year students study: Algorithms and Complexity, Computer Architectures, Maths for Computing, Object-Oriented Programming, Operating Systems, Problem Solving and Programming Fundamentals.",
-How will I be assessed?,"Each module can be assessed using different methods including in-class tests, coursework and exams. Most work is assessed by coursework.",
-What facilities are available?,"Students have access to computer labs and equipment including: Raspberry Pis, smartphones and virtual reality headsets.",
-What additional funding is available?,,University
-What if I want to do a gap year?,,
-How many hours of timetabled study is there each week?,,
-What do you look for in a personal statement?,,
-What are previous students doing after graduating?,,
-How much studying outside of contact hours am I expected to do each week?,,
-What happens if I don't quite meet the entry requirements?,,
-Will I have to do an interview before I am offered a place?,,
-What academic support is available?,,
-What on campus jobs are available?,,University
-What opportunities are available to develop employability skills?,,
-What connections do you have with local employers?,,
-How does doing a placement affect tuition fees?,,
-How much of the course is theory based and how much is practical?,,Course Overview
-What other qualifications are accepted?,,
-Do I have to do a final year project?,,
-What do I do if I have extenuating circumstances?,,University
-What support is available for students with disabilities?,,University
